refactor(client): modularize move queue geometry
